+++ Pandora/Paradigm/Inventory/Store.hs
@@ -4,12 +4,12 @@ module Pandora.Paradigm.Inventory.Store (Store (..), Storable, position, access,
 
 import Pandora.Core.Functor (type (:.), type (:=), type (<-|), type (~>))
 import Pandora.Core.Morphism ((%))
-import Pandora.Pattern.Category ((.), ($))
-import Pandora.Pattern.Functor.Covariant (Covariant ((<$>), (<$$>)))
+import Pandora.Pattern.Category (identity, (.), ($))
+import Pandora.Pattern.Functor.Covariant (Covariant ((<$>), (<$$>), (<$$$>)), (.|..))
 import Pandora.Pattern.Functor.Extractable (Extractable (extract))
-import Pandora.Pattern.Functor.Extendable (Extendable ((=>>)))
+import Pandora.Pattern.Functor.Extendable (Extendable ((=>>), (<<=$)))
 import Pandora.Pattern.Functor.Comonad (Comonad)
-import Pandora.Pattern.Functor.Adjoint ((|-))
+import Pandora.Pattern.Functor.Adjoint ((-|), (|-))
 import Pandora.Paradigm.Primary.Functor.Product (Product ((:*:)), type (:*:), attached)
 import Pandora.Paradigm.Controlflow.Effect.Adaptable (Adaptable (adapt))
 import Pandora.Paradigm.Controlflow.Effect.Interpreted (Interpreted (Primary, run))
@@ -20,14 +20,13 @@ import Pandora.Paradigm.Controlflow.Effect.Transformer.Comonadic (Comonadic (fli
 newtype Store p a = Store ((:*:) p :. (->) p := a)
 
 instance Covariant (Store p) where
-	f <$> Store (p :*: x) = Store . (:*:) p $ f <$> x
+	f <$> Store x = Store $ f <$$> x
 
 instance Extractable (Store p) where
-	extract (Store (p :*: f)) = f p
+	extract = (|- ($)) . run
 
 instance Extendable (Store p) where
-	Store (old :*: x) =>> f = Store . (:*:) old
-		$ \new -> f <$> Store $ new :*: x
+	Store x =>> f = Store $ f <$$> (Store .|.. (-| identity)) <$> x
 
 instance Comonad (Store p) where
 
@@ -44,14 +43,13 @@ instance Comonadic (Store p) where
 type Storable s x = Adaptable x (Store s)
 
 instance Covariant u => Covariant ((:*:) p <:<.>:> (->) p := u) where
-	f <$> TUT (p :*: x) = TUT . (:*:) p $ f <$$> x
+	f <$> TUT x = TUT $ f <$$$> x
 
 instance Extractable u => Extractable ((:*:) p <:<.>:> (->) p := u) where
 	extract = (|- extract) . run
 
 instance Extendable u => Extendable ((:*:) p <:<.>:> (->) p := u) where
-	TUT (old :*: x) =>> f = TUT . (:*:) old $ x =>> (\x' new -> f . TUT . (:*:) new $ x')
-
+	TUT x =>> f = TUT $ x <<=$ (\x' -> f . TUT . (x' -| identity))
 
 position :: Storable s t => t a -> s
 position = attached . run @(Store _) . adapt